remove settings to conftest
